Split TestWlcsDisplayServer out into header and implementation
